Improve sidebar - Added conditional rendering for the sidebar depending on the user viewing the page - Added buttons to hacker sidebar
