--- client/src/components/Sidebar.jsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-
-const drawerWidth = 240;
-
-const useStyles = makeStyles((theme) => ({
-  drawer: {
-    width: drawerWidth,
-  },
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  drawerContainer: {
-    overflow: 'auto',
-  },
-}));
-
-/**
- * Layout contains the Navbar and Sidebar that is present on all pages.
- */
-export default function Sidebar() {
-  const classes = useStyles();
-
-  return (
-    <Drawer
-      className={classes.drawer}
-      variant="permanent"
-      anchor="left"
-      classes={{
-        paper: classes.drawerPaper,
-      }}
-    >
-      <Toolbar />
-      <div className={classes.drawerContainer}>
-        <Typography variant="h5">Welcome, Hacker</Typography>
-      </div>
-    </Drawer>
-  );
-}
